# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a fundoo notes application.You can create notes here",
        "version": "1.0.0",
        "title": "Fundoo notes app",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/users/{userId}": {
            "get": {
                "tags": ["user"],
                "summary": "Get user by user id",
                "description": "",
                "operationId": "getUserById",
                "produces": [
                    "application/xml", "application/json"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The id that needs to be fetched. ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/User"
                        }
                    },
                    "400": {
                        "description": "Invalid userId supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": ["user"],
                "summary": "Updated user",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateUser",
                "produces": [
                    "application/xml", "application/json"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user id that need to be updated",
                        "required": true,
                        "type": "string"
                    }, {
                        "in": "body",
                        "name": "body",
                        "description": "Updated user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/User"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": ["user"],
                "summary": "Delete user",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteUser",
                "produces": [
                    "application/xml", "application/json"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The user id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/notes/{noteId}": {
          "get": {
              "tags": ["notes"],
              "summary": "Get note by note id",
              "description": "",
              "operationId": "getNoteById",
              "produces": [
                  "application/xml", "application/json"
              ],
              "parameters": [
                  {
                      "name": "noteId",
                      "in": "path",
                      "description": "The id that needs to be fetched. ",
                      "required": true,
                      "type": "string"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "$ref": "#/note"
                      }
                  },
                  "400": {
                      "description": "Invalid noteId supplied"
                  },
                  "404": {
                      "description": "note not found"
                  }
              }, "security": [
                {
                    "Bearer": []
                }
            ]

          },
          "put": {
            "tags": ["notes"],
            "summary": "Updated note",
            "description": "This can only be done by the logged in user.",
            "operationId": "updateNote",
            "produces": [
                "application/xml", "application/json"
            ],
            "parameters": [
                {
                    "name": "noteId",
                    "in": "path",
                    "description": "note id that need to be updated",
                    "required": true,
                    "type": "string"
                }, {
                    "in": "body",
                    "name": "body",
                    "description": "Updated note object",
                    "required": true,
                    "schema": {
                        "$ref": "#/note"
                    }
                }
            ],
            "responses": {
                "400": {
                    "description": "Invalid note supplied"
                },
                "404": {
                    "description": "note not found"
                }
            },
            "security": [
              {
                  "Bearer": []
              }
          ]
        },
        "delete": {
            "tags": ["notes"],
            "summary": "Delete note",
            "description": "This can only be done by the logged in user.",
            "operationId": "deleteNote",
            "produces": [
                "application/xml", "application/json"
            ],
            "parameters": [
                {
                    "name": "noteId",
                    "in": "path",
                    "description": "The note id that needs to be deleted",
                    "required": true,
                    "type": "string"
                }
            ],
            "responses": {
                "400": {
                    "description": "Invalid note id supplied"
                },
                "404": {
                    "description": "note not found"
                }
            },
            "security": [
              {
                  "Bearer": []
              }
          ]
        }

        }
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}